+++ test/smokes/pmd_java/expectations.rb
@@ -8,17 +8,6 @@ s.add_test(
   analyzer: { name: "PMD Java", version: default_version },
   issues: [
     {
-      message: "Do not add empty strings",
-      links: %W[https://pmd.github.io/pmd-#{default_version}/pmd_rules_java_performance.html#addemptystring],
-      id: "AddEmptyString",
-      path: "src/Hello.java",
-      location: { start_line: 17, start_column: 16, end_line: 17, end_column: 17 },
-      object: { ruleset: "Performance", priority: "3" },
-      git_blame_info: {
-        commit: :_, line_hash: "846dcdb29796bc0ef3e1c3b6c6ccd7ccd6c352f7", original_line: 17, final_line: 17
-      }
-    },
-    {
       message: "Avoid calling finalize() explicitly",
       links: %W[https://pmd.github.io/pmd-#{default_version}/pmd_rules_java_errorprone.html#avoidcallingfinalize],
       id: "AvoidCallingFinalize",
@@ -27,17 +16,6 @@ s.add_test(
       object: { ruleset: "Error Prone", priority: "3" },
       git_blame_info: {
         commit: :_, line_hash: "477c2ae2eb41ce1af10542ce1df1bcaf698305a8", original_line: 16, final_line: 16
-      }
-    },
-    {
-      message: "Use block level rather than method level synchronization",
-      links: %W[https://pmd.github.io/pmd-#{default_version}/pmd_rules_java_multithreading.html#avoidsynchronizedatmethodlevel],
-      id: "AvoidSynchronizedAtMethodLevel",
-      path: "src/Hello.java",
-      location: { start_line: 20, start_column: 18, end_line: 20, end_column: 29 },
-      object: { ruleset: "Multithreading", priority: "3" },
-      git_blame_info: {
-        commit: :_, line_hash: "928fb2851e5777de1b2700b5b9950bf745e62197", original_line: 20, final_line: 20
       }
     },
     {
@@ -60,17 +38,6 @@ s.add_test(
       object: { ruleset: "Security", priority: "3" },
       git_blame_info: {
         commit: :_, line_hash: "a7ec9e7adccc4fce30431f11e864f382c4d16e40", original_line: 12, final_line: 12
-      }
-    },
-    {
-      message: "System.out.println is used",
-      links: %W[https://pmd.github.io/pmd-#{default_version}/pmd_rules_java_bestpractices.html#systemprintln],
-      id: "SystemPrintln",
-      path: "src/Hello.java",
-      location: { start_line: 8, start_column: 9, end_line: 8, end_column: 26 },
-      object: { ruleset: "Best Practices", priority: "2" },
-      git_blame_info: {
-        commit: :_, line_hash: "addb4c033c4233fd7e9025796b0a22ff829782f9", original_line: 8, final_line: 8
       }
     }
   ],