+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 Misc:
 
 - **ESLint** Enable new recommended configuration [#2321](https://github.com/sider/runners/pull/2321)
+- **PMD Java** Enable new recommended configuration [#2345](https://github.com/sider/runners/pull/2345)
 - **RuboCop** Enable new recommended configuration [#2323](https://github.com/sider/runners/pull/2323)
 
 ## 0.48.2